requirements.txt
matplotlib
pyserial
PySide2